db creation via migrations, h2 db for tests
