+++ src/main/resources/db/migration/V1__setup.sql
@@ -1,0 +1,8 @@
+CREATE TABLE user (
+	id int(11) NOT NULL AUTO_INCREMENT,
+	login varchar(50) NOT NULL,
+	password varchar(50) NOT NULL,
+	max_score int(8) DEFAULT NULL,
+	PRIMARY KEY (id),
+	UNIQUE (login)
+);
