+++ src/test/java/ru/mail/park/java/controller/AbstractControllerTest.java
@@ -1,0 +1,20 @@
+package ru.mail.park.java.controller;
+
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.web.servlet.MockMvcPrint;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.transaction.annotation.Transactional;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+@AutoConfigureMockMvc(print = MockMvcPrint.NONE)
+@Transactional
+public abstract class AbstractControllerTest {
+
+	@Autowired
+	protected MockMvc mockMvc;
+}
