+++ src/test/resources/application.properties
@@ -1,9 +1,1 @@
 spring.main.banner-mode=off
-#spring.datasource.url=jdbc:mysql://localhost:3306/maze_db?useUnicode=true&characterEncoding=UTF-8
-#spring.datasource.username=root
-#spring.datasource.password=A0403959666a
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-
-spring.datasource.url=jdbc:mysql://nr84dudlpkazpylz.chr7pe7iynqr.eu-west-1.rds.amazonaws.com:3306/gn5ckts36aiep3vw?useSSL=false
-spring.datasource.username=v3mtyzmh0vz6oftf
-spring.datasource.password=l6uh0rhdq9o2szw8
